pyproject.toml
# Copyright 2022 Canonical Ltd.
# See LICENSE file for licensing details.

[tool.poetry]
package-mode = false
requires-poetry = ">=2.0.0"

[tool.poetry.dependencies]
python = "^3.10"
# breaking change in ops 2.10.0: https://github.com/canonical/operator/pull/1091#issuecomment-1888644075
ops = "^2.9.0, <2.10"
tenacity = "^9.1.2"
poetry-core = "^2.1.3"
jinja2 = "^3.1.6"
requests = "^2.32.4"
charm-refresh = "^3.1.0.0"

[tool.poetry.group.charm-libs.dependencies]
# data_platform_libs/v0/data_interfaces.py
ops = ">=2.0.0"
# tls_certificates_interface/v2/tls_certificates.py
# tls_certificates lib v2 uses a feature only available in cryptography >=42.0.5
cryptography = ">=42.0.5"
jsonschema = "*"
# grafana_agent/v0/cos_agent.py
pydantic = "*"
cosl = ">=0.0.50"
# tempo_coordinator_k8s/v0/charm_tracing.py
opentelemetry-exporter-otlp-proto-http = "1.21.0"

[tool.poetry.group.format]
optional = true

[tool.poetry.group.format.dependencies]
ruff = "^0.12.4"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
ruff = "^0.12.4"
codespell = "^2.4.1"

[tool.poetry.group.unit.dependencies]
pytest = "^8.4.1"
pytest-xdist = "^3.8.0"
pytest-cov = "^6.2.1"
ops-scenario = "^6.0.3, <6.0.4"  # 6.0.4 requires ops >= 2.12

[tool.poetry.group.integration.dependencies]
pytest = "^8.4.1"
pytest-operator = "^0.43.1"
juju = "^3.6.0.0"
mysql-connector-python = "~8.0.33"
tenacity = "^9.1.2"
ops = "^2.9.0"
allure-pytest-default-results = "^0.1.3"
tomli = "^2.2.1"
tomli-w = "^1.2.0"

[tool.poetry.group.build-refresh-version]
optional = true

[tool.poetry.group.build-refresh-version.dependencies]
charm-refresh-build-version = "^0.2.0"

[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true

[tool.pytest.ini_options]
minversion = "6.0"
log_cli_level = "INFO"

# Formatting tools configuration
[tool.ruff]
# preview and explicit preview are enabled for CPY001
preview = true
target-version = "py38"
src = ["src", "."]
line-length = 99

[tool.ruff.lint]
explicit-preview-rules = true
select = ["A", "E", "W", "F", "C", "N", "D", "I", "CPY001"]
ignore = [
    # Missing docstring in public method (pydocstyle doesn't look for docstrings in super class
    # https://github.com/PyCQA/pydocstyle/issues/309) TODO: add pylint check? https://github.com/PyCQA/pydocstyle/issues/309#issuecomment-1284142716
    "D102",
    "D105", # Missing docstring in magic method
    "D107", # Missing docstring in __init__
    "D403", # First word of the first line should be capitalized (false positive on "MySQL")
    "D415", # Docstring first line punctuation (doesn't make sense for properties)
    "E501", # Line too long (because using black creates errors with this)
    "N818", # Exception name should be named with an Error suffix
    "W505", # Doc line too long (so that strings in comments aren't split across lines)
]

[tool.ruff.lint.per-file-ignores]
# D100, D101, D102, D103: Ignore missing docstrings in tests
"tests/*" = ["D1"]

[tool.ruff.lint.flake8-copyright]
# Check for properly formatted copyright header in each file
author = "Canonical Ltd."
notice-rgx = "Copyright\\s\\d{4}([-,]\\d{4})*\\s+"

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pydocstyle]
convention = "google"
